pyproject.toml
[project]
name = "muos-app"
version = "0.0.1"
description = "A simple muOS app for RomM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "pillow>=11.1.0",
  "pip>=25.0.1",
  "pysdl2>=0.9.17",
  "python-dotenv>=1.1.0",
]
